faults fixes

+++ src/hu/elte/madtycoon/objects/Worker.java
@@ -5,14 +5,14 @@ import hu.elte.madtycoon.render.AnimatedSprite;
 import hu.elte.madtycoon.utils.Vector2F;
 
 public abstract class Worker extends Entity {
-    protected float salary;
+    protected int salary;
 
-    public Worker(World world, AnimatedSprite sprite, Vector2F position, float salary) {
+    public Worker(World world, AnimatedSprite sprite, Vector2F position, int salary) {
         super(world, sprite, position);
         this.salary = salary;
     }
 
-    public float getSalary() {
+    public int getSalary() {
         return salary;
     }
 }
